Add files via upload
A program that when run in a terminal, provides the user with the option to add toppings to their pizza, recognizing when the chosen topping is not available. At the end, the program asks the user if they want to clear all of the toppings from their pizza before making their final decision.